cleanup(web): delete ProfileEditor modal + drop /time route (#978)
Final cleanup pass for the Profile + Screen Time merge. With chart
restoration (#1036) and the full inline subsection chain (#1043 / #1045 /
#1053 / #1055 / #973) in place, every editable field on a profile is now
covered by an inline subsection on /profiles — the modal and the legacy
/time hub are both dead weight.

- Deleted the ProfileEditor modal and all of its callers/state in
  ProfilesPage. "+ New Profile" now opens a tiny inline name-only form;
  the new profile boots with safe-by-default body (LastKnownGood +
  Sum + no schedules / cap / blocked categories) and auto-expands so the
  operator fills the rest inline.
- Removed the /time route + TimePage + TimePage.test.tsx (no remaining
  imports). Remapped /time/:profileId/timeline → /profiles/:profileId/
  timeline; updated back-link copy. The Screen Time nav entry was already
  gone (#972); its placeholder comment in Layout.tsx is refreshed.
- Extracted formatMins / localBucketOffsetMin / groupBucketsByLocalDay
  to web/src/lib/timeFormat.ts so they survive TimePage's deletion.
  Re-pointed every importer (ProfilesPage, ProfileTimelineChart,
  DeviceTimelinePage). Util tests moved to a sibling test file.
- ProfilesPage.test.tsx: rewrote the create describe for the inline
  name-only flow (validation, success body, cancel, Enter/Escape).
  Dropped the failureMode-via-modal describe — those radios have no UI
  surface post-cleanup; the column default is still pinned by the create
  body assertion. Documented the orphan in code comments.
- net web/ diff: +394 / −1788 lines.

No router-side / API / shared/contract changes.
